Update to 5.18.0
This updates the package to the Linux kernel version 5.18.0,
together with updating the included Panfrost scheduler patch
according to the changes in the kernel.

This update also drops the previously included DP Alt Mode
patches, which was already described in [1].  In a few words,
those patches worked fine on some Pinebook Pros (PBPs), but
they also failed miserably on other PBPs, as reported by the
users.  The root cause is still unknown.

The included kernel configuration is a result of the automatic
configuration update for the new kernel version.  As such, it
introduces quite a few changes, but will not be used in that
form anyway.  The next commit will introduce a nearly completely
reworked kernel configuration, and all of the changes since
5.17.x will also be described there in detail.

[1] https://gitlab.manjaro.org/manjaro-arm/packages/core/linux-rc/-/merge_requests/3

Signed-off-by: Dragan Simic <dsimic@manjaro.org>