Disable CONFIG_INPUT_EVBUG
The "evbug" module basically tells the input subsystem
that it's a valid driver for all input devices, so it
gets loaded automatically when the first input device
is found, which is by no means the desired behavior.

The only reasonable solution is not to build the "evbug"
module.  Introducing some module blacklists would be
rather complicated and quite pointless.

Signed-off-by: Dragan Simic <dsimic@manjaro.org>